Move plugin usage hooks next to registry since they're part of its API
