+++ ui/plugin-system/src/components/PluginRegistry/plugin-registry-model.ts
@@ -12,6 +12,7 @@
 // limitations under the License.
 
 import { createContext, useContext } from 'react';
+import { useQuery, UseQueryOptions } from '@tanstack/react-query';
 import { PluginImplementation, PluginMetadata, PluginType } from '../../model';
 
 export interface PluginRegistryContextType {
@@ -21,10 +22,47 @@ export interface PluginRegistryContextType {
 
 export const PluginRegistryContext = createContext<PluginRegistryContextType | undefined>(undefined);
 
+/**
+ * Use the PluginRegistry context directly. This is meant as an escape hatch for custom async flows. You should probably
+ * be using `usePlugin` or `useListPluginMetadata` instead.
+ */
 export function usePluginRegistry() {
   const ctx = useContext(PluginRegistryContext);
   if (ctx === undefined) {
     throw new Error('PluginRegistryContext not found. Did you forget a provider?');
   }
   return ctx;
+}
+
+// Allows consumers to pass useQuery options from react-query when loading a plugin
+type UsePluginOptions<T extends PluginType> = Omit<
+  UseQueryOptions<PluginImplementation<T>, Error, PluginImplementation<T>, [string, PluginType, string]>,
+  'queryKey' | 'queryFn'
+>;
+
+/**
+ * Loads a plugin and returns the plugin implementation, along with loading/error state.
+ */
+export function usePlugin<T extends PluginType>(pluginType: T, kind: string, options?: UsePluginOptions<T>) {
+  // We never want to ask for a plugin when the kind isn't set yet, so disable those queries automatically
+  options = {
+    ...options,
+    enabled: (options?.enabled ?? true) && kind !== '',
+  };
+  const { getPlugin } = usePluginRegistry();
+  return useQuery(['getPlugin', pluginType, kind], () => getPlugin(pluginType, kind), options);
+}
+
+// Allow consumers to pass useQuery options from react-query when listing metadata
+type UseListPluginMetadataOptions = Omit<
+  UseQueryOptions<PluginMetadata[], Error, PluginMetadata[], [string, PluginType]>,
+  'queryKey' | 'queryFn'
+>;
+
+/**
+ * Gets a list of plugin metadata for the specified plugin type and returns it, along with loading/error state.
+ */
+export function useListPluginMetadata(pluginType: PluginType, options?: UseListPluginMetadataOptions) {
+  const { listPluginMetadata } = usePluginRegistry();
+  return useQuery(['listPluginMetadata', pluginType], () => listPluginMetadata(pluginType), options);
 }
