+++ ui/plugin-system/src/runtime/index.ts
@@ -12,7 +12,6 @@
 // limitations under the License.
 
 export * from './datasources';
-export * from './plugins';
 export * from './template-variables';
 export * from './time-range';
 export * from './time-series-queries';
